requirements.txt
streamlit
numpy
fpdf
qrcode
google-generativeai
